Refactor, move base stream workers to pixels-worker-common
