+++ pixels-turbo/pixels-worker-common/src/main/java/io/pixelsdb/pixels/worker/common/StreamWorkerCommon.java
@@ -17,7 +17,7 @@
  * License along with Pixels.  If not, see
  * <https://www.gnu.org/licenses/>.
  */
-package io.pixelsdb.pixels.worker.vhive;
+package io.pixelsdb.pixels.worker.common;
 
 import io.pixelsdb.pixels.common.physical.Storage;
 import io.pixelsdb.pixels.core.*;
@@ -26,7 +26,6 @@ import io.pixelsdb.pixels.core.reader.PixelsReaderOption;
 import io.pixelsdb.pixels.planner.plan.physical.domain.InputSplit;
 import io.pixelsdb.pixels.planner.plan.physical.domain.OutputInfo;
 import io.pixelsdb.pixels.planner.plan.physical.domain.StorageInfo;
-import io.pixelsdb.pixels.worker.common.WorkerCommon;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
