+++ pixels-turbo/pixels-worker-vhive/src/main/java/io/pixelsdb/pixels/worker/vhive/PartitionedJoinStreamWorker.java
@@ -22,7 +22,7 @@ package io.pixelsdb.pixels.worker.vhive;
 import io.pixelsdb.pixels.common.turbo.WorkerType;
 import io.pixelsdb.pixels.planner.plan.physical.input.PartitionedJoinInput;
 import io.pixelsdb.pixels.planner.plan.physical.output.JoinOutput;
-import io.pixelsdb.pixels.worker.common.BasePartitionedJoinWorker;
+import io.pixelsdb.pixels.worker.common.BasePartitionedJoinStreamWorker;
 import io.pixelsdb.pixels.worker.common.WorkerContext;
 import io.pixelsdb.pixels.worker.vhive.utils.RequestHandler;
